FrameMain and NavPanel's features.
#FrameMain
List object has been created to send to other class the original pointer of the main list with the objective of using only one list.

#NavPanel
Creation of the text label and field that we need to show de information.
Furthermore, it has created methods to change the pertinents label and field.  In the same way, this class contain the methods to enable o disable the update salary button.

#Note
We have to decide the name and features of a few new methods. !!